The pinned Mojo now rejects only 21 of the ordinary `assets/` `_ok` fixtures, down from 102, and all 21 are the single `__has_next__` iterator family that already has its own roadmap task. Thirty-five fixtures were respelled the way the pin demands — upstream's `Slice` descriptor, `Deinitable`, a hoisted recursion, `SIMD.lt(...)` masks, `UnsafePointer[T, MutUntrackedOrigin]`, `origin_of(self.field)` return origins, `ImplicitlyCopyable` on a frozen comptime value, `Sized` for `len`, concrete origins on the `RefBox` carriers, and a dozen more — and each is now a `conformance/cases.tsv` `run` row so the spelling cannot drift back. Twenty-two whose *shape* the pin refuses outright moved under `assets/extensions/`, where `scripts/sweep-assets-mojo --extensions` asserts the rejection and the whole pipeline still runs them: origin-parameter erasure and the exclusivity rules around it, cross-origin field stores, parametric-mut pointer writes, partial field moves, capturing-lambda locals, a `mut self` `__call__`, owned-interior generations, and the Mojito-only `std.algorithms` and `std.collections.string_dict`.
Five compiler changes came out of the respellings, each because upstream has a spelling Mojito lacked. `SIMD` gains the elementwise comparison methods `lt`/`le`/`gt`/`ge`/`eq`/`ne` on both backends, since upstream constrains infix `<` to `Scalar` and gives `==` whole-vector meaning; bool lanes order `False` below `True`, as the `i1` unsigned predicates do. `round` is nearest-ties-to-even (`llvm.roundeven.f64` natively), which also closes an entry on the arithmetic output-diff ledger. `origin_of(self.field)` is accepted wherever `origin_of(self)` is in an abstract signature, widening to the whole receiver rather than rejecting the upstream spelling. `Self.Ts[i]` folds like the bare `Ts[i]` pack index, and a bracketed projection index accepts a qualified struct binder (`Self.InnerType[Self.o]`).

`AGENTS.md` invariant 1 and `assets/README.md` now separate an extension *kept on purpose* — one tracking an announced upstream direction, listed in `docs/non-goals.md` — from a ledgered divergence Mojito means to withdraw; both keep their fixtures under `assets/extensions/`, and the second kind carries a header comment naming the divergence. `docs/roadmap.md` loses the closed checkbox and gains twenty-six divergence entries for what the pass found, from `Int / Int` truncating upstream and `SIMD.ne` against NaN answering neither IEEE predicate, down to three constructor leniencies and the `Writer` payload type.

+++ assets/ok/builtins_numeric.mojo
@@ -1,10 +1,11 @@
-# The utility numeric built-ins run identically on the tree-walker and the VM.
+# The utility numeric built-ins. `round` is ties-to-even, and a
+# `FloatLiteral` spells its truncation to `Int` through `Float64`.
 def main():
     print(abs(-5))
     print(abs(-3.5))
     print(min(3, 7), max(3, 7))
     print(min(2.5, 1.5))
     print(round(2.5), round(2.4))
-    print(Int(3.9), UInt(42), Float64(7))
+    print(Int(Float64(3.9)), UInt(42), Float64(7))
     var got: Int = abs(min(-4, -2))
     print(got)
